leaky in discriminator

+++ gan.py
@@ -34,10 +34,10 @@ class GAN():
                     # next MNIST batch
                     x_real, _= data.train.next_batch(BATCH_SIZE)
                     # because most values are 0
-                    x_real = 2 * (x_real - 0.5)
+                    x_real = 2 * (x_real.astype(np.float32) - 0.5)
                     
                     # random input to generator
-                    z = np.random.normal(0, 1, size=[BATCH_SIZE, DIM_Z])
+                    z = np.random.normal(0, 1, size=[BATCH_SIZE, DIM_Z]).astype(np.float32)
 
                     # get generator output
                     x_fake = sess.run(self.g.x_fake, feed_dict={self.g.z:z})
